数值AST从支持Int转变为支持Double

+++ Output/main.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
 extern "C" {
-    int average(int, int);
+    double average(double, double);
 }
 
 int main() {
-    std::cout << "average of 3.0 and 4.0: " << average(3, 4) << std::endl;
+    std::cout << "average of 3.0 and 4.0: " << average(3.0, 4.0) << std::endl;
 }
